feat: add traits for structs

+++ src/models/notifications.rs
@@ -1,7 +1,7 @@
 use bson::oid::ObjectId;
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, Hash)]
 #[serde(rename_all = "kebab-case")]
 pub enum NotificationType {
     Pin,
@@ -9,17 +9,17 @@ pub enum NotificationType {
     Normal,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq, Eq, Hash)]
 pub struct Notification {
     pub _id: ObjectId,
     pub title: String,
     pub content: String,
-    pub time: u64,
+    pub time: u128,
     #[serde(rename = "type")]
     pub notification_type: NotificationType,
     pub publisher: ObjectId,
     pub receivers: Option<Vec<ObjectId>>,
     pub anoymous: bool,
     pub global: bool,
-    pub expire: u64,
+    pub expire: u128,
 }
